Django-Middleware-0x03/requirements.txt
Django>=5.2.6
djangorestframework>=3.16.0
django-filter>=23.2
djangorestframework-simplejwt>=5.5.1

